can use with net_notify,download,loop mplayer

+++ fetch_and_play_video.rb
@@ -2,6 +2,7 @@
 #require "logger"
 require "open-uri"
 require "mail"
+require "gtk2"
 
 Mail.defaults do
   delivery_method :smtp, {
@@ -23,41 +24,63 @@ def mail_notification(msg)
       body msg
     end
   end
+  puts "notification mail sent!"
+rescue => e
+  puts e
+end
+
+def net_connected?
+  begin
+    true if open("http://www.xinyegroup.com/dalaoju.html")
+  rescue => e
+    puts e
+    mail_notification e.to_s
+  end
+end
+
+def clean_old_files
+  puts "clean all old files"
+  Dir.glob("*").each do |f|
+    system "rm -rf #{f}" unless f == "fetch_and_play_video.rb"
+  end
 end
 
 $last_video = nil
 $current_video = nil
 
+clean_old_files
+
 loop do
   begin
-    #video_path = open("/ljasdjfjadsf").read.scan
-    puts video_path = "http://younoter.com/assets/rails.png"
-    puts $current_video = video_name = video_path.strip.split("/").last
-    if $last_video and $current_video == $last_video
-      puts "无更新."
-    else
-      puts "已更新."
-      system "rm -rf #{$last_video}"
-      download = system "wget -p #{video_path}"
-      if download then
-        $last_video = $current_video
-        system "killall qiv"
-        system "qiv #{video_name} &"
+    if net_connected?
+      puts video_path = open("http://www.xinyegroup.com/dalaoju.html").read
+      #puts video_path = "http://younoter.com/assets/rails.png"
+      puts $current_video = video_name = video_path.strip.split("/").last
+      if $last_video and $current_video == $last_video
+        puts "无更新."
       else
-        puts "error"
-        raise "restart mplayer error!"
+        puts "已更新."
+        system "rm -rf #{$last_video}"
+        system "zenity --info --text='找到视频，正在努力下载，请确保网络正常...' &"
+        download = system "wget -c #{video_path}"
+        system "killall zenity"
+        if download then
+          $last_video = $current_video
+          system "killall mplayer"
+          system "mplayer -fs -loop 0 #{video_name} &"
+        else
+          puts "error"
+          raise "restart mplayer error!"
+        end
       end
+    else
+      #system "xmessage 网络连接异常... -timeout 3"
+      system "zenity --error --text='网络连接异常...' --timeout=3"
     end
   rescue => e
     puts e
-    mail_notification e
+    mail_notification e.to_s
   end
   sleep 3
   puts "检查是否有更新--->"
-end
-
-def clean_old_files
-  Dir.glob("*").each do |f|
-    system "rm -rf #{f}" unless if == "fetch_and_play_video.rb"
-  end
 end
